Changing file structure

+++ Scripts/Edit-UserName.ps1
@@ -1,70 +1,64 @@
+
 function Invoke-This {
     if (!([Security.Principal.WindowsPrincipal][Security.Principal.WindowsIdentity]::GetCurrent()).IsInRole([Security.Principal.WindowsBuiltInRole]"Administrator")) {
         Start-Process powershell.exe "-NoProfile -ExecutionPolicy Bypass -File `"$PSCommandPath`" $PSCommandArgs" -WorkingDirectory $pwd -Verb RunAs
         Exit
     }
     
-    $scriptName = "Add-LocalUser"
+    $scriptName = "Edit-UserName"
     $scriptPath = $env:TEMP
 
-    <#  if (Get-Content -Path "$PSScriptRoot\CS-Framework.ps1" -ErrorAction SilentlyContinue) {
+    if (Get-Content -Path "$PSScriptRoot\CS-Framework.ps1" -ErrorAction SilentlyContinue) {
         $framework = Get-Content -Path "$PSScriptRoot\CS-Framework.ps1" -Raw
-    } else { #>
-    Get-Script -Url "https://raw.githubusercontent.com/badsyntaxx/Chaste-Scripts/main/CS-Framework.ps1" -Target "$scriptPath\CS-Framework.ps1"
-    $framework = Get-Content -Path "$scriptPath\CS-Framework.ps1" -Raw
-    Get-Item -ErrorAction SilentlyContinue "$scriptPath\CS-Framework.ps1" | Remove-Item -ErrorAction SilentlyContinue
-    # }
+    } else {
+        Get-Script -Url "https://raw.githubusercontent.com/badsyntaxx/Chaste-Scripts/main/CS-Framework.ps1" -Target "$scriptPath\CS-Framework.ps1"
+        $framework = Get-Content -Path "$scriptPath\CS-Framework.ps1" -Raw
+        Get-Item -ErrorAction SilentlyContinue "$scriptPath\CS-Framework.ps1" | Remove-Item -ErrorAction SilentlyContinue
+    }
 
     $scriptDescription = @"
- This function creates a new local user account on a Windows system with specified settings, 
- including the username, optional password, and group. The account and password never expire.
+ This script allows you to modify the username of a user on a Windows system. 
 "@
 
     $core = @"
 function $scriptName {
     try {
         Get-Item -ErrorAction SilentlyContinue "$scriptPath\$scriptName.ps1" | Remove-Item -ErrorAction SilentlyContinue
-        Write-Host " Chaste Scripts: Add User v0315241122"
+        Write-Host " Chaste Scripts: Edit User Name v0315242300"
         Write-Host "$scriptDescription" -ForegroundColor DarkGray
 
-        Write-Text -Type "header" -Text "Enter name" -LineBefore -LineAfter
-        `$name = Get-Input -Prompt "" -Validate "^([a-zA-Z0-9 _\-]{1,64})$"  -CheckExistingUser
+        `$username = Select-User
 
-        Write-Text -Type "header" -Text "Enter password" -LineBefore -LineAfter
-        `$password = Get-Input -Prompt "" -IsSecure
-        
-        Write-Text -Type "header" -Text "Set group membership" -LineBefore -LineAfter
-        `$group = Get-Option -Options `$([ordered]@{
-            'Administrators' = 'Set this users group membership to administrators.'
-            'Users'          = 'Set this users group membership to standard users.' 
-        }) -ReturnValue -LineAfter
+        Write-Text -Type "header" -Text "Enter username" -LineBefore -LineAfter
+        `$newName = Get-Input -Prompt "" -Validate "^(\s*|[a-zA-Z0-9 _\-]{1,64})$" -CheckExistingUser
 
-        Write-Text -Type "notice" -Text "You're about to create a new local user!" -LineBefore -LineAfter
+        Write-Text -Type "notice" -Text "You're about to change this users name." -LineBefore -LineAfter
         `$choice = Get-Option -Options `$([ordered]@{
             "Submit"  = "Confirm and apply." 
             "Reset"   = "Start over at the beginning."
             "Exit"    = "Run a different command."
         }) -LineAfter
 
-        if (`$choice -ne 0 -and `$choice -ne 2) { Invoke-Script "$scriptName" }
-        if (`$choice -eq 2) {  Write-Exit -Script "$scriptName" }
+        if (`$choice -ne 0 -and `$choice -ne 1 -and `$choice -ne 2) { $scriptName }
+        if (`$choice -eq 1) { Invoke-Script "$scriptName" }
+        if (`$choice -eq 2) { Write-Exit -Script "$scriptName" }
+    
+        Rename-LocalUser -Name `$username -NewName `$newName
 
-        New-LocalUser `$name -Password `$password -Description "Local User" -AccountNeverExpires -PasswordNeverExpires -ErrorAction Stop | Out-Null
-        Add-LocalGroupMember -Group `$group -Member `$name -ErrorAction Stop
+        `$newUser = Get-LocalUser -Name `$newName
 
-        `$newUserName = Get-LocalUser -Name `$name | Select-Object -ExpandProperty Name
-        `$data = Get-AccountInfo `$newUserName
-
-        Write-Text -Type "list" -List `$data -LineAfter
-
-        if (`$null -ne `$newUserName) { Write-Exit -Message "The user account was created." -Script "$scriptName" } 
-        else { throw "There was an unknown error when creating the user." }
+        if (`$null -ne `$newUser) { 
+            `$data = Get-UserData `$newName
+            Write-Text -Type "list" -List `$data -LineAfter
+            Write-Exit "The name for this account has been changed." -Script "$scriptName"
+        } else {
+            Write-Text -Type "fail" -Text "There was an unknown error when trying to rename this user."
+        }
     } catch {
-        Write-Text -Type "error" -Text "Add user error: `$(`$_.Exception.Message)"
-        Write-Exit
+        Write-Text -Type "error" -Text "Edit name error: `$(`$_.Exception.Message)"
+        Write-Exit -Script "$scriptName"
     }
 }
- 
 
 "@
 
@@ -76,6 +70,7 @@ function $scriptName {
 
     Start-Process powershell.exe "-NoProfile -NoExit -ExecutionPolicy Bypass -File `"$scriptPath\$scriptName.ps1`"" -WorkingDirectory $pwd -Verb RunAs
 }
+
 function Get-Script {
     param (
         [Parameter(Mandatory)]
